👌 IMPROVE: Improve course list semantics
