+++ src/components/courses/index.jsx
@@ -1,4 +1,4 @@
-export { default as CourseItem } from 'components/courses/item'
 export { default as CourseList } from 'components/courses/CourseList'
 export { default as CourseSearch } from 'components/courses/CourseSearch'
+export { CourseItem } from 'components/courses/item'
 export { default as PageNo } from 'components/courses/PageNo'
